started clean up process
