--- Archived files/index.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<title>NiceBots - Control Panel</title>
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.0/jquery.min.js"></script>
-<script src="https://d3js.org/d3.v5.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.0.3/socket.io.js"></script>
-</head>
-<body>
-
-<h1 id="client_count">This is a Heading</h1>
-<p>This is a paragraph.</p>
-
-<script>
-    console.log("Script BEGIN");
-  var socket = io.connect('http://localhost');
-  console.log("Connection to server succeded");
-    socket.on('connect', function()
-    {
-        console.log("New connection");
-        socket.on('users_count', function(data)
-        {
-            console.log(data);
-            $('#client_count').text(data);
-            console.log("Connection");
-        });
-    });
-    console.log("Script END");
-</script>
-
-</body>
-</html>
