--- Leaderboards/Agreeableness.txt
@@ -1,8 +1,0 @@
-19 [charliekirk11]
-18 [realDonaldTrump]
-8 [pudofmuddle]
-8 [usreai]
-5 [joelrobinson_uk]
-5 [thanerosa]
-3 [desmund22]
-1 [disturbedboy69]
