--- Leaderboards/Conscientiousness.txt
@@ -1,8 +1,0 @@
-20 [charliekirk11]
-20 [realDonaldTrump]
-5 [pudofmuddle]
-5 [usreai]
-4 [disturbedboy69]
-2 [desmund22]
-1 [joelrobinson_uk]
-1 [thanerosa]
