--- Leaderboards/EmotionalRange.txt
@@ -1,8 +1,0 @@
-13 [disturbedboy69]
-9 [charliekirk11]
-9 [realDonaldTrump]
-6 [pudofmuddle]
-6 [usreai]
-2 [thanerosa]
-1 [desmund22]
-1 [joelrobinson_uk]
