--- Leaderboards/Extraversion.txt
@@ -1,8 +1,0 @@
-20 [charliekirk11]
-20 [realDonaldTrump]
-6 [usreai]
-4 [thanerosa]
-3 [desmund22]
-3 [joelrobinson_uk]
-3 [pudofmuddle]
-1 [disturbedboy69]
