--- Leaderboards/Openness.txt
@@ -1,8 +1,0 @@
-20 [realDonaldTrump]
-19 [charliekirk11]
-16 [joelrobinson_uk]
-10 [desmund22]
-10 [thanerosa]
-3 [disturbedboy69]
-2 [usreai]
-1 [pudofmuddle]
