--- Modules/Movement/LVL1.txt
@@ -1,3 +1,0 @@
-Faster than last place
-15
-Slow coach
