--- Modules/Movement/LVL2.txt
@@ -1,3 +1,0 @@
-Runaround
-25
-Speedy
