--- Modules/Movement/LVL3.txt
@@ -1,3 +1,0 @@
-Silver Bolt
-40
-Oh gosh I am so quick
