--- Modules/Power/LVL1.txt
@@ -1,3 +1,0 @@
-Accelerated Blast Modifier
-2
-My weapons fire!
