--- Modules/Power/LVL2.txt
@@ -1,3 +1,0 @@
-Super Accelerated Blast Modifier
-3
-Weapons shoot faster
