--- Modules/Power/LVL3.txt
@@ -1,3 +1,0 @@
-Relentless Attack Modifier
-4
-Spray and pray
