--- Modules/PowerUps/LVL1.txt
@@ -1,3 +1,0 @@
-PowerUp 1
-40
-This weapon is bla bla bla
