--- Modules/PowerUps/LVL2.txt
@@ -1,3 +1,0 @@
-Weapon 2
-60
-This weapon is bla bla bla
