--- Modules/Shields/LVL1.txt
@@ -1,3 +1,0 @@
-Mended Barrier
-200
-Some protection
