--- Modules/Shields/LVL2.txt
@@ -1,3 +1,0 @@
-Dooms Wall
-400
-I live longer than before
