--- Modules/Shields/LVL3.txt
@@ -1,3 +1,0 @@
-The Iron Maid
-600
-Titan health
