--- Modules/Weapons/LVL1.txt
@@ -1,3 +1,0 @@
-Close Defence Fusion Eraser
-40
-I deal damage
