--- Modules/Weapons/LVL2.txt
@@ -1,3 +1,0 @@
-Space Cannon
-60
-I deal more damage
