--- Modules/Weapons/LVL3.txt
@@ -1,3 +1,0 @@
-The Destructor
-80
-Wow my weapons are good
