--- Users/thanerosa/outputUnity.txt
@@ -1,5 +1,0 @@
-Openness 0.4557934228765461
-Conscientiousness 0.032047080314864806
-Extraversion 0.20039383762040025
-Agreeableness 0.0792915722166756
-Emotional range 0.16585617849130124
