+++ src/lib/utils/index.ts
@@ -1,0 +1,1 @@
+export const util = () => 'thing'
