+++ src/routes/fight/index.ts
@@ -1,0 +1,7 @@
+import express from 'express'
+
+const fight = express.Router()
+
+fight.post('/', (_, res) => res.send('Sample'))
+
+export { fight }
