+++ src/routes/index.ts
@@ -1,0 +1,13 @@
+import { ioRoute } from './ioRoute'
+import { root } from './root'
+import { user } from './user'
+import { fight } from './fight'
+import { leaderboard } from './leaderboard'
+
+export {
+    ioRoute,
+    root,
+    leaderboard,
+    user,
+    fight
+}
