+++ src/routes/leaderboard/index.ts
@@ -1,0 +1,8 @@
+import express from 'express'
+
+const leaderboard = express.Router()
+
+leaderboard.get('/', (_, res) => res.send('Sample'))
+leaderboard.post('/', (_, res) => res.send('Sample'))
+
+export { leaderboard }
