+++ src/routes/root/index.ts
@@ -1,0 +1,7 @@
+import express from 'express'
+
+const root = express.Router() 
+
+root.get('/', (_, res) => res.send('api root'))
+
+export { root }
