+++ src/routes/stats/index.ts
@@ -1,0 +1,7 @@
+import express from 'express'
+
+const stats = express.Router() 
+
+stats.get('/', (_, res) => res.send('api root'))
+
+export { stats }
